requirements.txt
# SuperHyperion Dependencies

# ============================================
# Core
# ============================================
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ============================================
# TypeDB
# ============================================
typedb-driver>=2.28.0

# ============================================
# LLM (Ollama)
# ============================================
ollama>=0.3.0
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.2.0

# ============================================
# CodeAct Sandbox
# ============================================
jupyter-client>=8.0.0
ipykernel>=6.0.0

# ============================================
# API
# ============================================
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sse-starlette>=2.0.0
python-multipart>=0.0.9

# ============================================
# UI
# ============================================
streamlit>=1.30.0
streamlit-agraph>=0.0.45
pyvis>=0.3.2

# ============================================
# Utilities
# ============================================
httpx>=0.27.0
numpy>=1.26.0
pandas>=2.2.0

# ============================================
# PDF Processing (for ingestion)
# ============================================
pymupdf>=1.24.0
pdfplumber>=0.11.0

# ============================================
# Testing
# ============================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
